requirements.txt
python-telegram-bot[webhooks]==20.7
google-generativeai==0.8.4
mysql-connector-python==8.2.0